Description
PostgreSQL allows columns of a table to be defined as variable-length multidimensional arrays. Arrays of any built-in or user-defined base type, enum type, or composite type can be created. Arrays of domains are not yet supported.To write an array value as a literal constant, enclose the element values within curly braces and separate them by commas.

We can also access arbitrary rectangular slices of an array, or subarrays. An array slice is denoted by writing lower-bound:upper-bound for one or more array dimensions. 

Refer:
https://www.postgresql.org/docs/9.1/static/arrays.html